requirements.txt
GPUtil
Pillow
PySimpleGUI
dlib
mediapipe
memory-profiler
mojimoji
nptyping
numpy
onnxruntime
opencv-python
protobuf
psutil
pyqrcode
requests
snakeviz
tqdm
typing_extensions
urllib3
onnx
onnxruntime
torch
torchvision
qrcode
ttkbootstrap
watchdog
faiss-gpu
scipy